+++ src/biosigpy/__init__.py
@@ -1,6 +1,7 @@
 """Python implementations of Biosiglib biomedical signal algorithms."""
 
 from biosigpy.ecg.pantompkins import pantompkins
+from biosigpy.ecg.sloperange import sloperange
 from biosigpy.hrv.tdmetrics import tdmetrics
 
-__all__ = ["pantompkins", "tdmetrics"]
+__all__ = ["pantompkins", "sloperange", "tdmetrics"]